+++ csrc/device_lower/pass/circular_buffer.cpp
@@ -9,7 +9,7 @@
 #include <ir/utils.h>
 #include <kernel_ir.h>
 
-#include <device_lower/pass/double_buffer.h>
+#include <device_lower/pass/circular_buffer.h>
 
 #include <algorithm>
 #include <iterator>
