+++ csrc/device_lower/pass/circular_buffer.h
